app/requirements.txt
streamlit
Pillow
torch
torchvision
pandas
roboflow
shapely
ultralytics

